Generalize `Plausible.Cache` warmers
